+++ pappl/printer.c
@@ -765,7 +765,7 @@ create_printer(
 #ifdef HAVE_STATFS
   if (statfs(system->directory, &spoolinfo))
     k_supported = INT_MAX;
-  else if ((spoolsize = (double)spoolinfo.f_bsize * spoolinfo.f_blocks / 1024) > INT_MAX)
+  else if ((spoolsize = (double)spoolinfo.f_bsize * (double)spoolinfo.f_blocks / 1024.0) > INT_MAX)
     k_supported = INT_MAX;
   else
     k_supported = (int)spoolsize;
@@ -773,7 +773,7 @@ create_printer(
 #elif defined(HAVE_STATVFS)
   if (statvfs(system->directory, &spoolinfo))
     k_supported = INT_MAX;
-  else if ((spoolsize = (double)spoolinfo.f_frsize * spoolinfo.f_blocks / 1024) > INT_MAX)
+  else if ((spoolsize = (double)spoolinfo.f_frsize * (double)spoolinfo.f_blocks / 1024.0) > INT_MAX)
     k_supported = INT_MAX;
   else
     k_supported = (int)spoolsize;
